make community providers an array of strings && move communites array to the profile identity
