+++ anchor/programs/pubkey-protocol/src/instructions/community/create_community.rs
@@ -10,7 +10,7 @@ pub struct CreateCommunity<'info> {
     #[account(
       init,
       payer = fee_payer,
-      space = Community::size(&[authority.key()], &[Identity { provider: PubKeyIdentityProvider::Solana, provider_id: authority.key().to_string(), name: "Community Creator Wallet".to_owned() }]),
+      space = Community::size(&[authority.key()], &[]),
       seeds = [
         PREFIX,
         COMMUNITY,
@@ -48,12 +48,6 @@ pub fn create_community(ctx: Context<CreateCommunity>, args: CreateCommunityArgs
         x,
     } = args;
 
-    let identity = Identity {
-        provider: PubKeyIdentityProvider::Solana,
-        provider_id: authority.key().to_string(),
-        name: "Community Creator Wallet".to_owned(),
-    };
-
     community.set_inner(Community {
         bump: ctx.bumps.community,
         slug,
@@ -62,7 +56,7 @@ pub fn create_community(ctx: Context<CreateCommunity>, args: CreateCommunityArgs
         fee_payers: vec![fee_payer],
         authority,
         pending_authority: None,
-        providers: vec![identity],
+        providers: vec![],
         discord,
         farcaster,
         github,
