+++ anchor/programs/pubkey-protocol/src/instructions/community/verify_community_profile.rs
@@ -9,7 +9,7 @@ pub struct VerifyProfileForCommunity<'info> {
     pub community: Account<'info, Community>,
     #[account(mut)]
     pub profile: Account<'info, Profile>,
-    #[account(constraint = authority.key() == community.authority @ PubkeyProfileError::UnauthorizedCommunityAction)]
+    #[account(constraint = community.check_for_authority(&authority.key()) @ PubkeyProfileError::UnAuthorized)]
     pub authority: Signer<'info>,
     #[account(mut)]
     pub fee_payer: Signer<'info>,
@@ -17,13 +17,11 @@ pub struct VerifyProfileForCommunity<'info> {
 }
 
 pub fn verify_community_profile(ctx: Context<VerifyProfileForCommunity>) -> Result<()> {
-    let community = &mut ctx.accounts.community;
-    let profile = &mut ctx.accounts.profile;
-    let authority = &ctx.accounts.authority;
+    // let community = &mut ctx.accounts.community;
+    // let profile = &mut ctx.accounts.profile;
+    // let authority = &ctx.accounts.authority;
 
-    community.add_profile_verification(profile.key(), authority.key())?;
-
-    profile.add_community_verification(community.key(), authority.key())?;
+    // community.add_profile_verification(profile.key(), authority.key())?;
 
     Ok(())
 }
