+++ anchor/programs/pubkey-protocol/src/state/mod.rs
@@ -2,8 +2,10 @@ pub mod community;
 pub mod identity;
 pub mod pointer;
 pub mod profile;
+pub mod provider;
 
 pub use community::*;
 pub use identity::*;
 pub use pointer::*;
 pub use profile::*;
+pub use provider::*;
